Componentize breadcrumbs
