+++ components/InfoBox/InfoBox.js
@@ -1,29 +1,10 @@
 import Image from 'next/image'
-import { Link } from 'react-router-i18n'
 import useInfoBox from '../../hooks/useInfoBox'
-import { useRouteMatch } from 'react-router-dom'
-import { startCase } from 'lodash'
 import classNames from 'classnames'
+import Breadcrumbs from './Breadcrumbs'
 
 const InfoBox = ({ title, children, breadcrumbs }) => {
   const { showInfoBox, toggleInfoBox } = useInfoBox()
-
-  const deriveBreadcrumbsFromUrl = () => {
-    const { url } = useRouteMatch()
-    const urlSections = url.split('/')
-    const derivedBreadcrumbs = []
-
-    urlSections.map((b, i, { length }) => {
-      // skip first and last (first is empty because it's the root, last is current page)
-      if (i !== 0 && i < length - 1)
-        derivedBreadcrumbs.push({ title: startCase(b), path: `/${b}` })
-    })
-    return derivedBreadcrumbs
-  }
-
-  const breadcrumbsToDisplay = breadcrumbs
-    ? breadcrumbs
-    : deriveBreadcrumbsFromUrl()
 
   return (
     <div
@@ -45,23 +26,7 @@ const InfoBox = ({ title, children, breadcrumbs }) => {
       <div className="w-full flex flex-col items-end justify-end md:justify-start h-auto max-h-90p">
         <div className="flex justify-between w-full p-4 md:px-0">
           <div className="flex flex-col items-start justify-start">
-            {breadcrumbsToDisplay &&
-              breadcrumbsToDisplay.length > 0 &&
-              breadcrumbsToDisplay.map((b) => {
-                return (
-                  <span className="flex flex-row items-center justify-start pb-1">
-                    <Link
-                      className="text-gray-600 font-sans font-semibold"
-                      to={b.path}
-                    >
-                      {b.title}
-                    </Link>
-                    <p className="text-gray-700 mx-2 my-0 font-black text-md font-sans">
-                      /
-                    </p>
-                  </span>
-                )
-              })}
+            <Breadcrumbs breadcrumbs={breadcrumbs} />
             <span className="text-white text-3xl font-semibold font-sans tracking-tight">
               {title}
             </span>
